requirements.txt
# Breast Cancer Histopathology Analysis - Dependencies
# For deployment on Render

# Core ML
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Gradio UI
gradio>=4.0.0

# Database
pymongo>=4.0.0

# Storage
requests>=2.28.0

# Environment
python-dotenv>=1.0.0
